+++ semanticdb/src/dotty/semanticdb/SemanticdbConsumer.scala
@@ -2,13 +2,12 @@ package dotty.semanticdb
 
 import scala.tasty.Reflection
 import scala.tasty.file.TastyConsumer
-import scala.tasty.util.TreeTraverser
 
 class SemanticdbConsumer extends TastyConsumer {
 
   final def apply(reflect: Reflection)(root: reflect.Tree): Unit = {
     import reflect._
-    object Traverser extends TreeTraverser[reflect.type](reflect) {
+    object Traverser extends TreeTraverser {
 
       override def traverseTree(tree: Tree)(implicit ctx: Context): Unit = tree match {
         case IsDefinition(tree) =>
