+++ tests/run-separate-compilation/tasty-extractors-3/quoted_1.scala
@@ -1,7 +1,6 @@
 import scala.quoted._
 
 import scala.tasty.Reflection
-import scala.tasty.util.TreeTraverser
 
 object Macros {
 
@@ -12,7 +11,7 @@ object Macros {
     import reflect._
 
     val buff = new StringBuilder
-    val traverser = new TreeTraverser(reflect) {
+    val traverser = new TreeTraverser {
       override def traverseTypeTree(tree: TypeOrBoundsTree)(implicit ctx: Context): Unit = {
         buff.append(tree.tpe.show)
         buff.append("\n\n")
