+++ tests/run-with-compiler/tasty-consumer/Test.scala
@@ -1,5 +1,4 @@
 import scala.tasty.Reflection
-import scala.tasty.util.TreeTraverser
 import scala.tasty.file._
 
 object Test {
@@ -12,7 +11,7 @@ class DBConsumer extends TastyConsumer {
 
   final def apply(reflect: Reflection)(root: reflect.Tree): Unit = {
     import reflect._
-    object Traverser extends TreeTraverser[reflect.type](reflect) {
+    object Traverser extends TreeTraverser {
 
       override def traverseTree(tree: Tree)(implicit ctx: Context): Unit = tree match {
         case IsDefinition(tree) =>
